Multiplatform Log Formating

+++ virtual_emulator.c
@@ -70,8 +70,8 @@ size_t eval_inst(size_t inst_address){
         for(size_t i = 0; i < vm.stack_size; i+=1){
             const Var var = vm.stack[i];
             printf(
-                "%lu-- i: %li, u: %lu, f: %f, ptr: %p\n",
-                i, var.as_int64, var.as_uint64, var.as_float64, var.as_ptr
+                "%" PRIu64 "-- i: %" PRId64 ", u: %" PRIu64 ", f: %f, ptr: %p\n",
+                (uint64_t)i, var.as_int64, var.as_uint64, var.as_float64, var.as_ptr
             );
         }
         return sizeof(Inst);
